Split articles in multiple sub-modules.
