+++ src/web/articles/mod.rs
@@ -1,0 +1,6 @@
+pub mod find;
+pub mod list;
+pub mod responses;
+
+pub use find::get_article;
+pub use list::list_articles;
